extract and flatten proof-lane execution
